Add the Netlify path rewrite bug workaround
https://answers.netlify.com/t/bug-in-non-trailing-slash-rewrite/452/24

Say we deploy an app to myapp.netlify.app, then rewrite
example.com/myapp to the deployed URL. The browser won't be able to
find any relative assets, because on Netlify (AFAIK),
example.com/myapp sees itself as myapp.html and not myapp/index.html.
If the user navigates to example.com/myapp/ (with the trailing slash)
instead, it works fine.

There shouldn't be a difference, but here we are.